Extract session-scoped state resets into a mixin (WS5 seam)
Move the 3 methods that clear session-scoped counters, compaction
progress, and per-turn placeholder-boundary bookkeeping on session
reset/rollover (_reset_session_counters, _reset_compaction_progress,
_reset_session_scoped_runtime_state) verbatim out of engine.py into a
new ResetStateMixin in reset_state.py. LCMEngine mixes it in before
ContextEngine, so `self` still resolves every reference; call sites,
shared state, and method names are unchanged.

Behaviour-neutral: the moved method bodies are byte-identical to main
(AST-slice compare, 0 diffs). ruff clean; full suite 1575 passed /
12 xfailed (unchanged from the v0.19.0 baseline).
